# src/utils/swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.1.0",
    "title": "CashBack Backend API",
    "contact": {
      "email": "naingoted@gmail.com"
    }
  },
  "host": "localhost:8000",
  "schemes": ["http"],
  "tags": [
    {
      "name": "Ruleset",
      "description": "Rule Sets"
    },
    {
      "name": "Transaction",
      "description": "Transaction data"
    },
    {
      "name": "Cashback",
      "description": "Cashback data"
    }
  ],
  "paths": {
    "/api/transaction": {
      "post": {
        "tags": ["Transaction"],
        "summary": "Create Transaction",
        "description": "",
        "operationId": "registerTransaction",
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Created Transaction",
            "required": true,
            "schema": {
              "$ref": "#/definitions/RegisterTransaction"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "created"
          }
        }
      }
    },
    "/api/ruleset": {
      "post": {
        "tags": ["Ruleset"],
        "summary": "Create Ruleset",
        "description": "",
        "operationId": "registerRuleset",
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Created Ruleset",
            "required": true,
            "schema": {
              "$ref": "#/definitions/RegisterRuleset"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "created"
          }
        }
      }
    },
    "/api/cashback": {
      "get": {
        "tags": ["Cashback"],
        "summary": "Get all avilable cashbacks",
        "description": "",
        "operationId": "getCashBacks",
        "produces": ["application/json"],
        "parameters": [],
        "responses": {
          "200": {
            "description": "ok"
          }
        }
      }
    }
  },

  "definitions": {
    "RegisterTransaction": {
      "type": "object",
      "properties": {
        "date": {
          "type": "string",
          "format": "date",
          "required": true
        },
        "customerId": {
          "type": "number",
          "required": true
        },
        "id": {
          "type": "number",
          "required": true
        }
      }
    },
    "RegisterRuleset": {
      "type": "object",
      "properties": {
        "startDate": {
          "type": "string",
          "format": "date",
          "required": true
        },
        "endDate": {
          "type": "string",
          "format": "date",
          "required": true
        },
        "redemptionLimit": {
          "type": "number"
        },
        "cashback": {
          "type": "number",
          "required": true
        },
        "budget": {
          "type": "number",
          "required": true
        },
        "minTransactions": {
          "type": "number"
        }
      }
    },
    "GetCashBacks": {
      "type": "object",
      "properties": {
        "is_request_success": {
          "type": "string"
        },
        "response_date_time": {
          "type": "string",
          "format": "date-time"
        },
        "data": {
          "type": "object"
        }
      }
    }
  },
  "externalDocs": {
    "description": "GitHub repository",
    "url": "https://github.com/naingoted/express-typescript-cashback"
  }
}